Add supports of many websockets frames in the same binary frame
This commit modify the way of processing binary entry frames.

Old way: taking binary frame as websocket frame (and bufferize if it was
not complete).

New way: it take a binary frame as one or many frames composing one or
many messages and bufferize if frames are not complete.

This new behavior makes autobahn execute new tests that was crashing
woketo. The fix is taking 4 bytes for the mask (sometimes it was less
because of a wrong usage of the BitManipulation class).

Switch to usage of generators