Non-respect of my copyright.
HI,

To begin, the archives do not keep the authors of the commits

Then in Github's source (https://help.github.com/articles/licensing-a-repository/) : "You're under no obligation to choose a license. However, without a license, the default copyright laws apply, meaning that you retain all rights to your source code and no one may reproduce, distribute, or create derivative works from your work. If you're creating an open source project, we strongly encourage you to include an open source license. The Open Source Guide provides additional guidance on choosing the correct license for your project." 

You choose a license without consulting me and denying my copyright. You have to respect that, these are not the same conditions as your Symfony argument. Argument not admissible. There was no license before the release of the 2nd version.

I am also entitled to refuse the MIT license since I am a co-author and there is no notion of quantity in relation to a license. 

So you're not just the sole owner of the project ! It does not work that way, I inquired.